requirements.txt
playwright
pytest
pytest-xdist
python-dotenv
allure-pytest
faker
pydantic